+++ packages/cocoa-discord/src/main/activity.ts
@@ -1,8 +1,9 @@
 import { ActivityType } from "discord-api-types/v10";
 import { ActivityOptions, Client } from "discord.js";
 
-import chalk from "chalk";
 import { readFile } from "node:fs/promises";
+
+import chalk from "chalk";
 
 import { ArrayLoader } from "./loader.js";