+++ packages/music-module/src/adapters/youtube.ts
@@ -7,6 +7,7 @@ import * as play from "play-dl";
 
 import { Playable } from "../core/types.js";
 import { beautifyNumber, parseLength, pickLast } from "../core/utils.js";
+import { getState, playNextMusicInQueue } from "../core/voice.js";
 
 /**
  * {@link Playable} that plays YouTube Video
@@ -103,5 +104,44 @@ export async function getYoutubeVideo(search: string) {
     return all_videos;
   } else {
     return "Unknown URL";
+  }
+}
+
+/**
+ * Add music to queue and play it if not playing
+ * @returns A Queueable object or string indicating failure reason
+ */
+export async function addMusicToQueue(
+  guildId: string,
+  search: string,
+  requester: string,
+) {
+  const videos = await getYoutubeVideo(search);
+
+  if (typeof videos === "string") {
+    return videos;
+  }
+
+  const state = getState(guildId);
+
+  if (Array.isArray(videos)) {
+    const toAdd = videos.map((v) => ({
+      data: new YoutubeAdapter(v),
+      requestedBy: requester,
+    }));
+
+    state.musicQueue.push(...toAdd);
+
+    if (!state.isPlaying) playNextMusicInQueue(guildId);
+    return toAdd[0]!;
+  } else {
+    const toAdd = {
+      data: new YoutubeAdapter(videos),
+      requestedBy: requester,
+    };
+
+    state.musicQueue.push(toAdd);
+    if (!state.isPlaying) playNextMusicInQueue(guildId);
+    return toAdd;
   }
 }